Extendable pipeline for frontend done
Pipeline can now build images by just inserting image name and directory with the Dockerfile
into the YAML file. If a "test" stage is present it runs the test. "test" stage
should not be the last stage in the Dockerfile. Rather, the last stage should be the production-ready
image definition.

Co-authored-by: Jonas Konrad <24313231+jalaka@users.noreply.github.com>
Co-authored-by: ValentinFutterer <80054001+ValentinFutterer@users.noreply.github.com>